+++ stream-chat-android-ai-assistant/src/main/kotlin/io/getstream/chat/android/ai/assistant/AiMessagesScreen.kt
@@ -19,6 +19,7 @@ package io.getstream.chat.android.ai.assistant
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.scrollBy
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
@@ -98,11 +99,14 @@ import io.getstream.chat.android.ui.common.state.messages.Reply
  * @param skipPushNotification If new messages should skip triggering a push notification when sent. False by default.
  * @param skipEnrichUrl If new messages being sent, or existing ones being updated should skip enriching the URL.
  * If URL is not enriched, it will not be displayed as a link attachment. False by default.
+ * @param verticalArrangement Vertical arrangement of the regular message list.
+ * Default: [Arrangement.Top].
  * @param threadMessagesStart Thread messages start at the bottom or top of the screen.
+ * Default: [ThreadMessagesStart.BOTTOM].
  * @param topBarContent custom top bar content to be displayed on top of the messages list.
  * @param bottomBarContent custom bottom bar content to be displayed at the bottom of the messages list.
- * Default: [ThreadMessagesStart.BOTTOM].
  */
+@Suppress("LongMethod")
 @Composable
 public fun AiMessagesScreen(
     viewModelFactory: MessagesViewModelFactory,
@@ -120,6 +124,7 @@ public fun AiMessagesScreen(
     onUserAvatarClick: (User) -> Unit = {},
     skipPushNotification: Boolean = false,
     skipEnrichUrl: Boolean = false,
+    verticalArrangement: Arrangement.Vertical = Arrangement.Top,
     threadMessagesStart: ThreadMessagesStart = ThreadMessagesStart.BOTTOM,
     aiStartButton: @Composable BoxScope.() -> Unit = {
         DefaultAiStartButton(
@@ -234,6 +239,7 @@ public fun AiMessagesScreen(
                 viewModel = listViewModel,
                 reactionSorting = reactionSorting,
                 messagesLazyListState = state,
+                verticalArrangement = verticalArrangement,
                 threadMessagesStart = threadMessagesStart,
                 onThreadClick = remember(composerViewModel, listViewModel) {
                     {