+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/messages/MessageListTest.kt
@@ -16,6 +16,7 @@
 
 package io.getstream.chat.android.compose.ui.messages
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -65,6 +66,26 @@ internal class MessageListTest : SnapshotTest {
                     isLoading = true,
                 ),
                 reactionSorting = ReactionSortingByCount,
+            )
+        }
+    }
+
+    @Test
+    fun `loaded top aligned messages`() {
+        snapshotWithDarkMode {
+            MessageList(
+                messageListState = TwoMessagesListState,
+                verticalArrangement = Arrangement.Top,
+            )
+        }
+    }
+
+    @Test
+    fun `loaded bottom aligned messages`() {
+        snapshotWithDarkMode {
+            MessageList(
+                messageListState = TwoMessagesListState,
+                verticalArrangement = Arrangement.Bottom,
             )
         }
     }
@@ -206,14 +227,31 @@ private val LoadedMessageListState = MessageListState(
     ),
 )
 
+private val TwoMessagesListState = MessageListState(
+    messageItems = listOf(
+        MessageItemState(
+            message = PreviewMessageData.message3,
+            isMine = true,
+            showMessageFooter = true,
+        ),
+        MessageItemState(
+            message = PreviewMessageData.message5,
+            isMine = false,
+            showMessageFooter = true,
+        ),
+    ),
+)
+
 @Composable
 private fun MessageList(
     messagesLazyListState: MessagesLazyListState = rememberMessagesLazyListState(),
+    verticalArrangement: Arrangement.Vertical = Arrangement.Top,
     messageListState: MessageListState,
 ) {
     MessageList(
         messagesLazyListState = messagesLazyListState,
         currentState = messageListState,
         reactionSorting = ReactionSortingByCount,
+        verticalArrangement = verticalArrangement,
     )
 }